Uses the latest lists for endpoint
I have spoken to engineers in the Core team and they mentioned that this
endpoint we were using is old and not their preferred route for this
api.

They believe that once FT.com stops using it they might be able to
remove it, this is a step towards that.